+++ packages/stylelint-config-wordpress/CHANGELOG.md
@@ -1,3 +1,7 @@
+# HEAD
+
+-   Changed: Updated `stylelint` peer dependency version to `^8.0.0`.
+
 # 12.0.0
 
 -   Changed: `stylelint-config-wordpress` now extends [`stylelint-config-recommended`](https://github.com/stylelint/stylelint-config-recommended), which turns on the `at-rule-no-unknown`, `block-no-empty`, `comment-no-empty`, `declaration-block-no-ignored-properties`, `declaration-block-no-redundant-longhand-properties`, `font-family-no-duplicate-names`, `media-feature-name-no-unknown`, `no-empty-source` rule. These rules are part of stylelint's [possible errors](https://github.com/stylelint/stylelint/blob/master/docs/user-guide/rules.md#possible-errors) rules.