+++ packages/stylelint-config-wordpress/CHANGELOG.md
@@ -6,6 +6,7 @@
 -   Removed: Jest snapshots.
 -   Removed: `stylelint < 9.1.3` compatibility.
 -   Updated: `stylelint-config-recommended` to `2.1.0`.
+-   Updated: `stylelint-scss` to `2.1.0`.
 -   Updated: Bump minimum Node.js required version to `8.9.3`.
 
 # 12.0.0